requirements.txt
Flask
Flask-SQLAlchemy 
Flask-Migrate 
Flask-Script
pg8000 
mailjet_rest
python-magic
rarfile
pyjwt
gunicorn
pebble
humanfriendly